add FarmSecret, SessionExData and support version 2.9.364

+++ VpnHood.AccessServer.Test/Dom/SessionDom.cs
@@ -15,7 +15,7 @@ public class SessionDom
     public AccessToken AccessToken { get; }
     public SessionRequestEx SessionRequestEx { get; }
     public SessionResponseEx SessionResponseEx { get; private set; }
-    public long SessionId => SessionResponseEx.SessionId;
+    public long SessionId => (long)SessionResponseEx.SessionId;
 
     private SessionDom(TestInit testInit, AgentClient agentClient, AccessToken accessToken, SessionRequestEx sessionRequestEx, SessionResponseEx sessionResponseEx)
     {
